fix: add gender and rename package
